Move arm_model_evaluator.py to util folder
